Fixed user auth
Previously, I stored the id_token as the user_id in the database. The
token is supposed to be sent to the backend and validated to obtain the
user_id. Added validate_user function to main.py which returns the
user_id based on a token. The tokens change from session to session, so
it is not a reliable way of keeping track of a user.